+++ CHANGES.md
@@ -3,7 +3,7 @@
 
 Changes in 13.0
 ---------------
-- Add `&modfunc=save` to form in Courses.php
+- Add `&modfunc=save` to form in Courses.php & Assignments.php
 - Security: SQL prevent INSERT or UPDATE on any table in Courses.php
 - Add "Block" mention after Period title in Courses.php
 - Add Course Period Form inputs action hook in Courses.php & Actions.php